Handle losing the session better.
- Added a function to force the login state
- Both CollectionBase and ModelBase force the login state
  to false on a 401 from a fetch.

- Added testHelps/modelTools for tests.  This prevents a
  problem where mocks weren't working because a module
  was imported before mocking occured.

- Restyled Top.scss to look a little better.
- Changed the utility background to be less busy.
