+++ src/Mode/Song/SongList.test.js
@@ -9,6 +9,8 @@ jest.mock('../../Network/Login');
 import * as Network from '../../Network/Network';
 import useIntersectionObserver, * as mockObserver from '../../Hooks/useIntersectionObserver';
 
+import { makeModels, makeAModel } from '../../testHelpers/modelTools';
+
 import BombayLoginContext from '../../Context/BombayLoginContext';
 import BombayUtilityContext from '../../Context/BombayUtilityContext';
