+++ src/Mode/Top.scss
@@ -21,4 +21,13 @@
             }
         }
     }
+
+    >div {
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-start;
+        align-items: center;
+        gap: 5px;
+        margin-top: 10px;
+    }
 }
