+++ src/Model/CollectionBase.test.js
@@ -15,6 +15,8 @@ jest.mock('../Network/Login');
 
 import * as Network from "../Network/Network";
 
+import { makeModels, makeAModel } from '../testHelpers/modelTools';
+
 import ModelBase from './ModelBase';
 import CollectionBase from './CollectionBase';
