+++ src/Model/SongModel.test.js
@@ -1,5 +1,7 @@
 import * as Network from "../Network/Network";
 
+import { makeModels, makeAModel } from '../testHelpers/modelTools';
+
 import ModelBase from './ModelBase';
 import SongModel from './SongModel';
