+++ src/Widgets/PickerList.test.js
@@ -8,6 +8,8 @@ jest.mock('../Network/Login');
 import * as Network from '../Network/Network';
 import * as mockObserver from '../Hooks/useIntersectionObserver';
 
+import { makeModels } from '../testHelpers/modelTools';
+
 import BombayLoginContext from '../Context/BombayLoginContext';
 
 import ModelBase from '../model/ModelBase';
