+++ src/global.scss
@@ -15,7 +15,7 @@ $standard-font-size: 20px;
 $button-font-size: $standard-font-size + 10px;
 
 @mixin utiliity-component-background {
-    background-image: radial-gradient(rgba(127, 0, 0, 0.7), rgba(0, 127, 0, 0.7), rgba(0, 0, 127, 0.7));
+    background-image: radial-gradient(rgba(15, 15, 15, 0.7), rgba(63, 63, 63, 0.7), rgba(127, 127, 127, 0.7));
 
     a {
         color: white;
